Patch [ 906210 ] from Josh Giles: SMUX connections are closed by the SNMP daemon
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V4-2-patches@9747 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/smux/smux.c
@@ -1517,7 +1517,7 @@ smux_trap_process(u_char *rsp, size_t *len)
 {
 	oid sa_enterpriseoid[MAX_OID_LEN], var_name[MAX_OID_LEN];
 	size_t datalen, var_name_len, var_val_len, maxlen;
-	int sa_enterpriseoid_len;
+	size_t sa_enterpriseoid_len;
 	u_char vartype, *ptr, *var_val;
 
 	long trap, specific;
